Remove duplication in CSV concerns
